Separate useReactEffectActions
