+++ packages/xstate-inspect/examples/server.ts
@@ -1,6 +1,6 @@
 import * as WebSocket from 'ws';
 import { createMachine, interpret, send } from 'xstate';
-import { toSCXMLEvent } from 'xstate/src/utils';
+import { toSCXMLEvent } from 'xstate/lib/utils';
 import { inspect } from '@xstate/inspect/lib/server';
 
 inspect({
