+++ packages/xstate-react/src/useMachine.ts
@@ -1,5 +1,4 @@
-import { useState, useEffect, useRef } from 'react';
-import useIsomorphicLayoutEffect from 'use-isomorphic-layout-effect';
+import { useState, useEffect } from 'react';
 import {
   EventObject,
   StateMachine,
@@ -18,7 +17,6 @@ import {
   ReactEffectType
 } from './types';
 import { useInterpret } from './useInterpret';
-import { partition } from './utils';
 
 function createReactActionFunction<TContext, TEvent extends EventObject>(
   exec: ActionFunction<TContext, TEvent>,
@@ -56,7 +54,7 @@ export type ActionStateTuple<TContext, TEvent extends EventObject> = [
   State<TContext, TEvent, any, any>
 ];
 
-function executeEffect<TContext, TEvent extends EventObject>(
+export function executeEffect<TContext, TEvent extends EventObject>(
   action: ReactActionObject<TContext, TEvent>,
   state: State<TContext, TEvent, any, any>
 ): void {
@@ -120,83 +118,6 @@ export function useMachine<
       ? State.create(options.state)
       : initialState) as State<TContext, TEvent, any, TTypestate>;
   });
-
-  const effectActionsRef = useRef<
-    Array<
-      [
-        ReactActionObject<TContext, TEvent>,
-        State<TContext, TEvent, any, TTypestate>
-      ]
-    >
-  >([]);
-  const layoutEffectActionsRef = useRef<
-    Array<
-      [
-        ReactActionObject<TContext, TEvent>,
-        State<TContext, TEvent, any, TTypestate>
-      ]
-    >
-  >([]);
-
-  useIsomorphicLayoutEffect(() => {
-    const sub = service.subscribe((currentState) => {
-      if (currentState.actions.length) {
-        const reactEffectActions = currentState.actions.filter(
-          (action): action is ReactActionObject<TContext, TEvent> => {
-            return (
-              typeof action.exec === 'function' &&
-              '__effect' in (action as ReactActionObject<TContext, TEvent>).exec
-            );
-          }
-        );
-
-        const [effectActions, layoutEffectActions] = partition(
-          reactEffectActions,
-          (action): action is ReactActionObject<TContext, TEvent> => {
-            return action.exec.__effect === ReactEffectType.Effect;
-          }
-        );
-
-        effectActionsRef.current.push(
-          ...effectActions.map<ActionStateTuple<TContext, TEvent>>(
-            (effectAction) => [effectAction, currentState]
-          )
-        );
-
-        layoutEffectActionsRef.current.push(
-          ...layoutEffectActions.map<ActionStateTuple<TContext, TEvent>>(
-            (layoutEffectAction) => [layoutEffectAction, currentState]
-          )
-        );
-      }
-    });
-
-    return () => {
-      sub.unsubscribe();
-    };
-  }, []);
-
-  // this is somewhat weird - this should always be flushed within useLayoutEffect
-  // but we don't want to receive warnings about useLayoutEffect being used on the server
-  // so we have to use `useIsomorphicLayoutEffect` to silence those warnings
-  useIsomorphicLayoutEffect(() => {
-    while (layoutEffectActionsRef.current.length) {
-      const [
-        layoutEffectAction,
-        effectState
-      ] = layoutEffectActionsRef.current.shift()!;
-
-      executeEffect(layoutEffectAction, effectState);
-    }
-  }, [state]); // https://github.com/davidkpiano/xstate/pull/1202#discussion_r429677773
-
-  useEffect(() => {
-    while (effectActionsRef.current.length) {
-      const [effectAction, effectState] = effectActionsRef.current.shift()!;
-
-      executeEffect(effectAction, effectState);
-    }
-  }, [state]);
 
   return [state, service.send, service];
 }
